feat: DeleteController returns 400 when id is missing

+++ src/adapters/presentation/controllers/delete-book/delete-book-controller.ts
@@ -2,7 +2,8 @@ import { DeleteBookUseCase } from "../../../../usecases/delete-book/delete-book-
 import { BaseController } from "../http/base-controller";
 import { HttpRequest } from "../http/http-request";
 import { HttpResponse } from "../http/http-response";
-import { deletedWithNoDescription } from '../helpers/http-helper'
+import { deletedWithNoDescription, badRequest } from '../helpers/http-helper'
+import { MissingParameter } from "../http/errors/missing-parameter";
 export class DeleteBookController implements BaseController {
     private bookRepository: DeleteBookUseCase
 
@@ -11,8 +12,16 @@ export class DeleteBookController implements BaseController {
     }
 
     async handle(httpRequest: HttpRequest): Promise<HttpResponse> {
-        const result = await this.bookRepository.deleteBook(httpRequest.body)
-        return new Promise(resolve => resolve(deletedWithNoDescription()))
+        try {
+            if(httpRequest.body === undefined) {
+                throw new MissingParameter('id')
+            }
+            const result = await this.bookRepository.deleteBook(httpRequest.body)
+            return new Promise(resolve => resolve(deletedWithNoDescription()))
+        } catch(error) {
+            return new Promise((resolve, reject) => reject(badRequest(error)))
+        }
+    
     }
     
 }
